+++ src/Errors/NoCurrentTaskFoundError.ts
@@ -1,0 +1,5 @@
+export class NoCurrentTaskFoundError extends Error {
+  constructor() {
+    super('No current task found')
+  }
+}